Make core work kind of
