+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/execution/sam/SamDebugSupport.kt
@@ -9,22 +9,13 @@ import com.intellij.execution.process.ProcessOutputTypes
 import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.execution.ui.ConsoleView
 import com.intellij.execution.ui.ConsoleViewContentType
-import com.intellij.openapi.application.ExpirableExecutor
-import com.intellij.openapi.application.impl.coroutineDispatchingContext
-import com.intellij.openapi.application.runInEdt
 import com.intellij.openapi.util.Key
 import com.intellij.openapi.util.registry.Registry
-import com.intellij.util.concurrency.AppExecutorUtil
 import com.intellij.xdebugger.XDebugProcessStarter
-import kotlinx.coroutines.launch
-import org.jetbrains.concurrency.AsyncPromise
-import org.jetbrains.concurrency.Promise
 import software.aws.toolkits.core.utils.getLogger
-import software.aws.toolkits.core.utils.warn
-import software.aws.toolkits.jetbrains.utils.ApplicationThreadPoolScope
-import software.aws.toolkits.resources.message
-import java.util.Timer
-import kotlin.concurrent.schedule
+import software.aws.toolkits.jetbrains.services.lambda.steps.AttachDebugger
+import software.aws.toolkits.jetbrains.utils.execution.steps.Context
+import software.aws.toolkits.jetbrains.utils.execution.steps.Step
 
 interface SamDebugSupport {
     fun numberOfDebugPorts(): Int = 1
@@ -36,48 +27,13 @@ interface SamDebugSupport {
      */
     fun samArguments(debugPorts: List<Int>): List<String> = emptyList()
 
-    fun createDebugProcessAsync(
-        environment: ExecutionEnvironment,
-        state: SamRunningState,
-        debugHost: String,
-        debugPorts: List<Int>
-    ): Promise<XDebugProcessStarter> {
-        val promise = AsyncPromise<XDebugProcessStarter>()
-        val bgContext = ExpirableExecutor.on(AppExecutorUtil.getAppExecutorService()).expireWith(environment).coroutineDispatchingContext()
-
-        val timerTask = Timer("Debugger Worker launch timer", true).schedule(debuggerConnectTimeoutMs()) {
-            if (!promise.isDone) {
-                runInEdt {
-                    promise.setError(message("lambda.debug.process.start.timeout"))
-                }
-            }
-        }
-
-        ApplicationThreadPoolScope(environment.runProfile.name).launch(bgContext) {
-            try {
-                val debugProcess = createDebugProcess(environment, state, debugHost, debugPorts)
-
-                runInEdt {
-                    promise.setResult(debugProcess)
-                }
-            } catch (t: Throwable) {
-                LOG.warn(t) { "Failed to start debugger" }
-                runInEdt {
-                    promise.setError(t)
-                }
-            } finally {
-                timerTask.cancel()
-            }
-        }
-
-        return promise
-    }
+    fun additionalDebugProcessSteps(environment: ExecutionEnvironment, state: SamRunningState): List<Step> = listOf()
 
     suspend fun createDebugProcess(
         environment: ExecutionEnvironment,
-        state: SamRunningState,
         debugHost: String,
-        debugPorts: List<Int>
+        debugPorts: List<Int>,
+        context: Context
     ): XDebugProcessStarter
 
     companion object {
