+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/java/JavaRuntimeDebugSupport.kt
@@ -6,13 +6,6 @@ package software.aws.toolkits.jetbrains.services.lambda.java
 import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.xdebugger.XDebugProcessStarter
 import software.aws.toolkits.jetbrains.services.lambda.execution.sam.RuntimeDebugSupport
-import software.aws.toolkits.jetbrains.services.lambda.execution.sam.SamRunningState
+import software.aws.toolkits.jetbrains.utils.execution.steps.Context
 
-class JavaRuntimeDebugSupport : RuntimeDebugSupport {
-    override suspend fun createDebugProcess(
-        environment: ExecutionEnvironment,
-        state: SamRunningState,
-        debugHost: String,
-        debugPorts: List<Int>
-    ): XDebugProcessStarter = JavaDebugUtils.createDebugProcess(environment, state, debugHost, debugPorts)
-}
+
